requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
pandas>=2.1.0
openpyxl>=3.1.0
click>=8.1.0
